+++ Sail-a-dex/Patches/EventPatches.cs
@@ -1,7 +1,6 @@
 ﻿using HarmonyLib;
 using System.Linq;
 using UnityEngine;
-using SailwindModdingHelper;
 
 namespace sailadex
 {