+++ Sail-a-dex/Utilities/Names.cs
@@ -96,8 +96,8 @@ namespace sailadex
             "Sen'na"
         };
 
-        public static string[] capitals = 
-        { 
+        public static string[] capitals =
+        {
             "grc", "dc", "fa", "kb"
         };
 
@@ -128,7 +128,8 @@ namespace sailadex
             "TimesSlept",
             "StormsWeathered",
             "FlotsamEncounters",
-            "SeaLifeEncounters"
+            "DenseFogEncounters",
+            "SeaLifeEncounters",
         };
 
         public static string[] transitNames =
@@ -147,5 +148,69 @@ namespace sailadex
             "KbFa"
         };
 
+        public static string[] regions =
+        {
+            "Aa", "Ea", "Ae", "Ffl",
+        };
+
+        public static string[] regionTransitNames =
+        {
+            "AaEa",
+            "AaAe",
+            "AaFfl",
+            "EaAa",
+            "EaAe",
+            "EaFfl",
+            "AeAa",
+            "AeEa",
+            "AeFfl",
+            "FflAa",
+            "FflEa",
+            "FflAe",
+        };
+
+        public static string[] alankhIslands =
+        {
+            "island 1 A (gold rock)",
+            "island 2 A (AlNilem)",
+            "island 3 A (Neverdin)",
+            "island 4 A (fish island)",
+            "island 5 A (clear mind)",
+            "island 6 A (lion fang)",
+            "island 7 A (alchemist's island)",
+            "island 8 A (academy)",
+            "island 20 A (Oasis)",
+        };
+
+        public static string[] emeraldIslands =
+        {
+            "island 9 E (dragon cliffs)",
+            "island 10 E (sanctuary)",
+            "island 11 E (crab beach)",
+            "island 12 E (New Port)",
+            "island 13 E (Sage Hills)",
+            "island 22 E (serpent isle)",
+        };
+
+        public static string[] mediIslands =
+        {
+            "island 15 M (Fort)",
+            "island 16 M (Sunspire)",
+            "island 17 M (Mount Malefic)",
+            "island 18 M (HappyBay)",
+            "island 19 M (Eastwind)",
+            "island 21 M (Siren Song)",
+            "island 23 M oracle",
+            "island 25 (chronos)",
+        };
+
+        public static string[] lagoonIslands =
+        {
+            "island 26 Lagoon Temple",
+            "island 27 Lagoon Shipyard",
+            "island 28 Lagoon Senna",
+            "island 29 Lagoon Onna",
+            "island 31 Lagoon Fisherman",
+        };
     }
 }